fix(ui): list account decks in the lobby and vs-AI pickers
The deck store persists only decks without an accountDeckId, so a
signed-in player's collection lives on the server. The lobby picker,
the vs-AI picker and the lobby preselection still read savedDecks
alone, so they showed nothing until the deck was opened in the editor.

Add useOwnedDecks, the merge PlayDeckShelf already did inline, and read
it from every surface that lists or resolves an owned deck.
