+++ src/AGENTS.md
@@ -45,7 +45,7 @@ Read first: `/AGENTS.md`, `docs/STYLE_GUIDELINES.md`, `docs/agents/UI_THEME_RULE
 - **One page-padding system.** Page-level horizontal padding is always `px-4 sm:px-6 lg:px-8` (the same scale the `TopBar` uses) — no per-view `px` inventions and no centered `max-w-* mx-auto` page boxes; inner text measures (`max-w-2xl` etc.) are fine.
 - **One exported component per file.** Files past ~200 lines split.
 - **No new abstractions for one-off patterns.** Three similar lines beat a helper.
-- **Accounts are optional and never gate play.** `useAuthStore` holds the bearer session (token persisted, `hydrate()` runs once from `AppShell`); the global `SignInDialog` (mounted in `AppShell`, opened via `stores/useSignInDialogStore`) handles GitHub/Discord OAuth, the email code flow, and first-run handle claim; `/auth/callback` (`views/AuthCallback.tsx`) receives OAuth redirects and non-consuming magic links. Full-page sign-in stores a short-lived return route and sanitized pending publication in `lib/authReturn.ts`, so callbacks restore the initiating page and action, including links opened in a new tab. Signed-out My Decks saves remain browser-local. `useLocalDeckAccountSync` uploads those local decks after sign-in and removes each local entry only after its account create succeeds; failed entries remain recoverable. Signed-in My Decks renders one collection backed by the account and creates immutable versions on later saves. The top-bar account button (`TopBar`) opens the dialog or, when signed in, Preferences. Relay usernames below are a separate, unrelated identity for now (#467).
+- **Accounts are optional and never gate play.** `useAuthStore` holds the bearer session (token persisted, `hydrate()` runs once from `AppShell`); the global `SignInDialog` (mounted in `AppShell`, opened via `stores/useSignInDialogStore`) handles GitHub/Discord OAuth, the email code flow, and first-run handle claim; `/auth/callback` (`views/AuthCallback.tsx`) receives OAuth redirects and non-consuming magic links. Full-page sign-in stores a short-lived return route and sanitized pending publication in `lib/authReturn.ts`, so callbacks restore the initiating page and action, including links opened in a new tab. Signed-out My Decks saves remain browser-local. `useLocalDeckAccountSync` uploads those local decks after sign-in and removes each local entry only after its account create succeeds; failed entries remain recoverable. Signed-in My Decks renders one collection backed by the account and creates immutable versions on later saves. `useDeckStore` persists only decks without an `accountDeckId`, so `savedDecks` alone is never the player's collection: every surface that lists or resolves an owned deck reads `hooks/useOwnedDecks.ts`, which merges account decks over the local ones. The top-bar account button (`TopBar`) opens the dialog or, when signed in, Preferences. Relay usernames below are a separate, unrelated identity for now (#467).
 - **Deck ownership and publication are separate.** Account decks are canonical editable resources exposed by `/api/decks`; immutable version snapshots back their cards and history. Publishing creates a separate `/api/deckhub/entries` record tied to one selected version, and a canonical deck may have multiple publications. Shipped presets are ownerless canonical decks with official publications; the first signed-in use creates one idempotent account-owned fork that preserves preset provenance. Tags and favorites belong to public entries. Top Decks uses named, dated ranking snapshots containing published entries only, so every visible ranking can open its exact card snapshot. Ranking evidence lives in the Hub database and new relay evidence uses a dedicated channel; analytics is only a one-time historical import source. Telemetry is never presented as a deck publication.
 - **Preset format compatibility.** The shipped preset JSON catalog still contains legacy `historicBrawl` values even though the shared `DeckFormat` protocol and selectable format registry expose `brawl`; both frontend expansion and Hub ingestion normalize those presets to `brawl`. Keep that normalization until the source catalog is migrated atomically across web, desktop, and Hub consumers.
 - **Presets are per-engine.** `stores/usePresetDecksStore` is the single load path (main-thread fetch of `public/preset_decks/`; the engine worker no longer serves presets). A preset JSON may declare `engines: EngineKind[]` (`lib/presetDecks.ts`); a preset without the field predates per-engine curation and is treated as playable everywhere except Ironsmith, whose card pool is verified per deck. Every picker that starts a game filters through `usePresetDecks(engine)` / `presetSupportsEngine` with the engine it will launch on (offline: `resolveOfflineEngine`; lobby: the room's `engine`); browse-only surfaces (deck editor) list all presets; the play-home grid hides presets whose only engines are unavailable (Ironsmith needs the Settings toggle) and shows `EngineMark` badges on engine-specific tiles.
